fix: split RefreshError into own file to satisfy max-classes-per-file
ESLint max-classes-per-file rule disallows having both RefreshError
and TokenRefreshManager in the same file. Move RefreshError and
classifyError to RefreshError.ts, re-export from TokenRefreshManager.ts.
